Switch to uint16_t for operatorReward and pay all to owner if payee not set
This allows to specify operator rewards between 0% and 100% in 0.01%
increments. Also makes setting the scriptOperatorPayout optional and
defaults to paying the full reward to the owner.

+++ src/evo/deterministicmns.h
@@ -113,7 +113,7 @@ public:
 
     uint256 proTxHash;
     uint32_t nCollateralIndex;
-    uint8_t operatorReward;
+    uint16_t operatorReward;
     CDeterministicMNStateCPtr state;
 
 public:
